adding logo muni
