+++ apps/financial-aid/web-veita/src/components/LogoMuni/Logo.treat.ts
@@ -1,0 +1,7 @@
+import { style } from 'treat'
+import { theme } from '@island.is/island-ui/theme'
+
+export const logo = style({
+  width: theme.spacing[6],
+  height: theme.spacing[6],
+})
