+++ apps/financial-aid/web-veita/src/components/LogoMuni/LogoSvg.tsx
@@ -1,0 +1,24 @@
+import React from 'react'
+import cn from 'classnames'
+
+import * as styles from './Logo.treat'
+
+interface LogoSvgProps {
+  name: string
+  className?: string
+}
+
+const LogoSvg = ({ name, className }: LogoSvgProps) => {
+  return (
+    <div
+      className={cn({
+        [`${styles.logo}`]: true,
+        [`${className}`]: className,
+      })}
+    >
+      <img src={`./svg/${name ? name : 'hfj'}.svg`} alt="" />
+    </div>
+  )
+}
+
+export default LogoSvg
