+++ apps/financial-aid/web-veita/src/components/LogoMuni/index.tsx
@@ -2,12 +2,15 @@ import React from 'react'
 import cn from 'classnames'
 
 import LogoSvg from './LogoSvg'
+import { useRouter } from 'next/router'
 
 interface LogoProps {
   className?: string
 }
 
 const LogoHfj = ({ className }: LogoProps) => {
+  const router = useRouter()
+
   return (
     <a
       href="https://www.hafnarfjordur.is/"
@@ -15,7 +18,7 @@ const LogoHfj = ({ className }: LogoProps) => {
       rel="noopener noreferrer"
       className={cn({ [`${className}`]: true })}
     >
-      <LogoSvg name="hfj.svg" />
+      <LogoSvg name={router.query.sveitarfelag as string} />
     </a>
   )
 }
